Separate like dislike functionality for Story to its own controller and service
