+++ app/Http/Controllers/CommentLikeDislikeController.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class CommentLikeDislikeController extends Controller
+{
+  public function getLikeDislike()
+  {
+  }
+
+  public function createLikeDislike()
+  {
+  }
+
+  public function deleteLikeDislike()
+  {
+  }
+}
